App completed in TSX: All functionalities implemented and converted to TypeScript
