+++ env.d.ts
@@ -1,0 +1,3 @@
+declare module '@env' {
+    export const API_KEY: "94fb27c0c12d4873cf9ce9afc8887574";
+  }
