+++ package.json
@@ -11,11 +11,12 @@
   "dependencies": {
     "@react-native-picker/picker": "^2.9.0",
     "@react-navigation/bottom-tabs": "^7.0.7",
-    "@react-navigation/native": "^7.0.4",
-    "@react-navigation/stack": "^7.0.6",
+    "@react-navigation/native": "^7.0.8",
+    "@react-navigation/stack": "^7.0.13",
     "@reduxjs/toolkit": "^2.3.0",
     "axios": "^1.7.7",
     "expo": "~52.0.11",
+    "expo-linear-gradient": "~14.0.1",
     "expo-location": "~18.0.2",
     "expo-status-bar": "~2.0.0",
     "react": "18.3.1",
@@ -28,8 +29,7 @@
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.1.0",
     "react-native-vector-icons": "^10.2.0",
-    "react-redux": "^9.1.2",
-    "expo-linear-gradient": "~14.0.1"
+    "react-redux": "^9.1.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
@@ -37,7 +37,13 @@
     "@babel/plugin-transform-private-methods": "^7.25.9",
     "@babel/plugin-transform-private-property-in-object": "^7.25.9",
     "@react-native-community/cli": "^15.1.2",
-    "metro-react-native-babel-preset": "^0.77.0"
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@types/react-native": "^0.73.0",
+    "@types/react-native-dotenv": "^0.2.2",
+    "@types/react-native-vector-icons": "^6.4.18",
+    "metro-react-native-babel-preset": "^0.77.0",
+    "typescript": "^5.7.2"
   },
   "private": true
 }
